test parity between java and scala

+++ faunadb-java/src/main/java/com/faunadb/client/java/query/Paginate.java
@@ -6,15 +6,15 @@ import com.google.common.base.Optional;
 @JsonSerialize(using=Codec.PaginateSerializer.class)
 public class Paginate implements Identifier {
   public static Paginate create(Identifier resource) {
-    return new Paginate(resource, Optional.<Long>absent(), Optional.<Cursor>absent(), Optional.<Long>absent());
+    return new Paginate(resource, Optional.<Long>absent(), Optional.<Cursor>absent(), Optional.<Integer>absent());
   }
 
   private final Identifier resource;
   private final Optional<Long> ts;
   private final Optional<Cursor> cursor;
-  private final Optional<Long> size;
+  private final Optional<Integer> size;
 
-  Paginate(Identifier resource, Optional<Long> ts, Optional<Cursor> cursor, Optional<Long> size) {
+  Paginate(Identifier resource, Optional<Long> ts, Optional<Cursor> cursor, Optional<Integer> size) {
     this.resource = resource;
     this.ts = ts;
     this.cursor = cursor;
@@ -29,7 +29,7 @@ public class Paginate implements Identifier {
     return new Paginate(resource, ts, Optional.of(cursor), size);
   }
 
-  public Paginate withSize(Long size) {
+  public Paginate withSize(Integer size) {
     return new Paginate(resource, ts, cursor, Optional.of(size));
   }
 
@@ -45,7 +45,7 @@ public class Paginate implements Identifier {
     return cursor;
   }
 
-  public Optional<Long> size() {
+  public Optional<Integer> size() {
     return size;
   }
 }
